requirements.txt
Django==3.2
djangorestframework==3.12.4
marshmallow==3.11.1
